requirements.txt
tensorflow
pillow
scipy
IPython
opencv-python
numpy